Define mask only inside bit_read()
The mask static constant is only ever used in the bit_read() inline
function. But (maybe because of a bug), GCC emits it even if the
inline is not used (and thus not emitted).

Moving the variable inside the function makes gcc aware that its use
is linked to the other and doesn't emit it when not needed.