Make it possible to disable examples building.
Most distributions will not install examples for their users, so
building them is just a waste of time during their build process.

For source distributions like Gentoo, FreeBSD's Ports or pkgsrc, this
means a waste of time for users installing the package.

On the other hand, during development, examples can be used to make
sure that the library is still in a state where it can be used, so
disabling them entirely might not be what is wanted.

This patch adds a configure switch to explicitly disable building of
examples during 'make all', still leaving an option to build them with
an explicit request.