Use AC_CANONICAL_HOST rather than AC_CANONICAL_SYSTEM.
The AC_CANONICAL_SYSTEM macro (deprecated in favour of
AC_CANONICAL_TARGET in recent versions of autoconf) is used for
packages that have both an host (system they run on) and a target
(system they run for), like (cross-)compilers, analysers, debuggers,
and pkg-config.

User applications and libraries usually only have an host.